jbogenturfa'i: compile the Lojban morphology and grammar.
I define a dependency on my morphology and header file in the
generation of the source file jbogerna.scm.  This file is then
included in the library as the grammar for Lojban, and used
in the definition of the parser.

As well, define the jbogenturfahi program, which uses this library
to create a command-line program to parser Lojban.

Note that this is not yet compilling for me, as I run out of memory
compiling the .c file generated for this egg.

git-svn-id: https://code.call-cc.org/svn/chicken-eggs/release/4/jbogenturfahi/trunk@21687 fca3e652-9b03-0410-8d7b-ac86a6ce46c4