jbogenturfa'i: remove unneeded rules in MORPHOLOGY INTERFACE section.
These rules are designed to work with the cmaxes parser, which
constructs a tree using the morphology file and then reparses that
tree when constructing the grammar.

jbogenturfa'i will take a more integrated approach, and will accept
the tree as it comes from the morphology file, without reparsing the
result from that file.

git-svn-id: https://code.call-cc.org/svn/chicken-eggs/release/4/jbogenturfahi/trunk@22185 fca3e652-9b03-0410-8d7b-ac86a6ce46c4