jbogenturfa'i: knr on gerna.peg
This is the first of a series of patches I'll be applying to clean
up the formatting of the grammar file.  The file is parsed by a perl
program in camxes, rendering it somewhat ugly for my working with
it.

I'm not done making the file prettier and easy to work with, but
this is a first pass.

git-svn-id: https://code.call-cc.org/svn/chicken-eggs/release/4/jbogenturfahi/trunk@22187 fca3e652-9b03-0410-8d7b-ac86a6ce46c4