jbogenturfa'i: #!key match all ? and * operators in morphology.
By default, I now ignore any ? and * productions that do not produce
any results.  I want this behavior in the grammar, but for the
morphology I want to preserve the old behavior of generating an
empty string for ? and an empty list for *.

I do this by converting all ? and * productions in the morphology
file to be #!key arguments, for which I can supply a default value.

The grammar generated by jbogenturfa'i is now significantly cleaner.



git-svn-id: https://code.call-cc.org/svn/chicken-eggs/release/4/jbogenturfahi/trunk@22349 fca3e652-9b03-0410-8d7b-ac86a6ce46c4