jbogenturfa'i: create code stubs for grammar.
This is a huge commit, representing most (but definitely not all) of
the work required to make nice parse trees come out of the Lojban
grammar.

I believe I've finished stubbing out all of the grammar productions,
but much of the code that is called needs to be cleaned up to
generate good parse trees.

I couldn't avoid making corrections to the grammar while making
these changes, so some grammar cleanup is also present here.

git-svn-id: https://code.call-cc.org/svn/chicken-eggs/release/4/jbogenturfahi/trunk@22594 fca3e652-9b03-0410-8d7b-ac86a6ce46c4