jbogenturfa'i: add ZOI and CLL.19.10 tests.
This is my first set of test cases.  I had to implement a large part
of the grammar to make these work, though they themselves are very
small part of the overall test suite.

I started with ZOI because it was the only rule where I required
extra-peg information, and so I needed to test it early anyway.

git-svn-id: https://code.call-cc.org/svn/chicken-eggs/release/4/jbogenturfahi/trunk@22596 fca3e652-9b03-0410-8d7b-ac86a6ce46c4