jbogenturfa'i: add bu, si, and zo tests.
I need to checkpoint this code, as I've found several problems in
the grammar relating to Magic Words.  I need to pause to fix these,
so check in what I've got--comment out the few productions currently
affected and let the rest remain as test cases as I modify the
grammar.

git-svn-id: https://code.call-cc.org/svn/chicken-eggs/release/4/jbogenturfahi/trunk@22785 fca3e652-9b03-0410-8d7b-ac86a6ce46c4